feat: rename filedata provider to regimer provider
