+++ src/ui/main.tsx
@@ -4,7 +4,7 @@ import routes from "./router/routes.tsx";
 import { UserSettingsProvider } from "./services/UserSettingsProvider.tsx";
 import { ThemeProvider } from "./services/ThemeProvider.tsx";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { FileDataProvider } from "./services/FileDataProvider.tsx";
+import { RegimeProvider } from "./services/RegimeProvider.tsx";
 import { MenuProvider } from "./services/MenuProvider.tsx";
 
 import "@mantine/core/styles.css";
@@ -19,14 +19,14 @@ createRoot(document.getElementById("root")!).render(
     <QueryClientProvider client={queryClient}>
         <ThemeProvider>
             <UserSettingsProvider>
-                <FileDataProvider>
+                <RegimeProvider>
                     <MenuProvider
                         isDev={env.isDev}
                         navigate={(path) => router.navigate(path)}
                     >
                         <RouterProvider router={router} />
                     </MenuProvider>
-                </FileDataProvider>
+                </RegimeProvider>
             </UserSettingsProvider>
         </ThemeProvider>
     </QueryClientProvider>,
